backend/pyproject.toml
[project]
name = "bitcoach"
version = "0.1.0"
description = "AI-powered trading coach for Upbit spot traders"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "anthropic>=0.40.0",
    "langgraph>=0.2.0",
    "langchain-anthropic>=0.3.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "pyjwt>=2.9.0",
    "pandas>=2.2.0",
    "sse-starlette>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "httpx>=0.27.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"
